+++ docker/lost-base/mamba_env.sh
@@ -1,5 +1,5 @@
 mamba create -n lost imagesize mysqlclient pandas numpy pytest adlfs \
     dask distributed bokeh flask==1.1.2 flask-jwt-extended==3.13.1 pyjwt=1.7.1 \
     flask-ldap3-login flask-mail flask-restx flask-sqlalchemy flask-user \
-    flask-wtf fsspec opencv \
+    flask-wtf fsspec opencv juypterlab \
     pudb python-igraph sphinx sk-video sqlalchemy flask-cors scikit-image psycopg2 pyarrow