+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.0] - unreleased
 ## Added
 - Install script
+- JupyterLab for Admins
 ## Changed
 - Use fsspec for filesystem abstraction
   * SIA: Do not send any image urls to frontend